requirements.txt
# Core backend dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# Sarvam AI for Indian language TTS/STT
sarvamai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# File uploads and async operations
aiofiles>=23.0.0
python-multipart>=0.0.6

# Ollama integration for local LLM
ollama>=0.1.7
